Large commit - ongoing process Many changes in files but some parts is NOT done yet, will make smaller commits in the future with more structure of the process.
I fixed some parts of the login-functionallity. Now you can login, and see
the stats only if you have a password. Though all the navigation-states
are not done.

All the files including the test, overview and final form (client) are
not done yet, still some issues to solve. Alsoe poll-API on the server is
not done.